Conflicts:
	Sunset/Sunset/SettingsViewController.m

+++ Sunset/Sunset/ViewController.m
@@ -102,7 +102,7 @@
 }
 
 - (void)setNotifications {
-  [sunEventObject setNotificationsWithSeconds: 60 * [[myDefaults objectForKey:@"notificationTimeCustomization"] integerValue]];
+  [sunEventObject setNotificationsWithSeconds: (int) (60 * [[myDefaults objectForKey:@"notificationTimeCustomization"] integerValue])];
 }
 
 - (void)viewDidLoad {
